+++ Documentation/CHANGES.md
@@ -14,6 +14,7 @@ See Unstable branch until now.
 
 - Pass.app: release 0.2. Added Edit button and action to the List Panel.
 - Launcher.app: release 0.2. Full French localization (help bundle, UI and Info Panel)
+- AgnostepManager: release 0.3. Full French localization (help bundle, UI and Info Panel)
 
 ### Changes of the Release Candidate 2.0.4